[OV][Tests] Merge test_fq_params_calculation.py and test_calculation_quantizer_params.py
